requirements.txt
requests==2.9.1
cryptography==1.2.3
setuptools==20.7.0
Django==1.11.17
itsdangerous==0.24
djangorestframework==3.9.2
